refactor(split-rendering): extract subsystems (Phase 2 partial)
Extract four more self-contained modules:

- base_tokens.rs: build_base_tokens, build_base_tokens_binary + classifiers
- transforms.rs: apply_wrapping_transform, apply_soft_breaks,
  apply_conceal_ranges, inject_virtual_lines
- folding.rs: apply_folding, fold_adjusted_visible_count,
  fold_indicators_for_viewport, diff_indicators_for_viewport + FoldIndicator
- scrollbar.rs: render_scrollbar, render_horizontal_scrollbar,
  render_composite_scrollbar, scrollbar_line_counts,
  scrollbar_visual_row_counts, compute_max_line_length

None depend on any shared render-time mega struct.
All 37 split_rendering tests pass.
